comparison between GFD.90 and Bliss job package
The similarities column depicts the color relevant to the methods
between methods from gfd and saga. Red means that they are extremely
dissimilar (eg. parameters are vastly different or method functionality
is nowgere near similar) or it does not exist in both gfd and bliss.
Light green means there is notable similarity. On the gfd side, the same
applies, except orange shows dissimilarities in parameters in methods.